+++ AndroidStudioProjects/SecondProjectByMVVM2/app/src/main/java/com/example/secondprojectbymvvm/view/mealitemlist/category/CategoryAdapter.kt
@@ -75,7 +75,7 @@ class CategoryAdapter(
             favoriteButton.setOnClickListener {
                 val favoriteMealName = meal.strMeal
                 val favoriteMealId = meal.idMeal
-                val userId = sharedPreferences.getInt(LoginActivity.USER_ID, 0)
+                val userId = sharedPreferences.getLong(LoginActivity.USER_ID, 0).toInt()
                 val favoriteMealPicture = meal.strMealThumb
                 val favorite = Favorite(
                     favoriteMealId.toLong(),
